test(nexus): move intelligence route tests out of the server/api tree
Five Vitest files sat beside their handlers under server/api, which
check:api-routes reserves for runtime route handlers, so the guard had been
failing since the commit that added them.

They move to the established test/api hierarchy with their relative references
rewritten to the convention already used there (../../../server/utils/... back
out to the app root). Nothing else in them changes, so their positive, auth,
quota, stream and sanitization coverage is unchanged.

The guard is not weakened; it now passes because server/api holds only handlers.
